all-requirements.txt
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.2
black==23.3.0
build==0.10.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-debug-toolbar-force==0.2
django-environ==0.10.0
django-filter==23.2
django-nine==0.2.7
django-stubs==4.2.1
django-stubs-ext==4.2.1
djangorestframework==3.14.0
djangorestframework-stubs==3.14.1
exceptiongroup==1.1.1
flake8==6.0.0
flake8-bugbear==23.5.9
flake8-builtins==2.1.0
flake8-commas==2.1.0
flake8-import-order==0.18.2
flake8-isort==6.0.0
flake8-print==5.0.0
flake8-variables-names==0.0.5
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pep8-naming==0.13.3
Pillow==9.5.0
pip-tools==6.13.0
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pyproject_hooks==1.0.0
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
